Fold nested_continue case into do_while test
Move the nested-loop-continue example into test/do_while as a new
`nested_continue` function instead of a standalone test folder, keeping
all do-while desugaring cases together.
