--- test/do_while_nested_continue/Makefile
@@ -1,1 +1,0 @@
-../_templates/Makefile
