--- test/do_while_nested_continue/fstar.fst.config.json
@@ -1,1 +1,0 @@
-../_templates/fstar.fst.config.json
