--- test/do_while_nested_continue/pal.config.json
@@ -1,1 +1,0 @@
-../_templates/pal.config.json
