--- test/do_while_nested_continue/pal.h
@@ -1,1 +1,0 @@
-../pal.h
